requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-heroku==0.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
